LPS-135718 Move utils/string to data-engine-js-components-web
- https://issues.liferay.com/browse/LPS-135718
