+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/core/reducers/pageValidationReducer.es.js
@@ -12,8 +12,7 @@
  * details.
  */
 
-import {sub} from 'dynamic-data-mapping-form-field-type/util/strings.es';
-
+import {sub} from '../../utils/strings';
 import {PagesVisitor} from '../../utils/visitors.es';
 import {EVENT_TYPES} from '../actions/eventTypes.es';
