+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/index.es.js
@@ -30,6 +30,7 @@ export {
 } from './utils/repeatable.es';
 export * as RulesSupport from './utils/rulesSupport';
 export {default as setDataRecord} from './utils/setDataRecord.es';
+export * as StringUtils from './utils/strings';
 export {PagesVisitor, RulesVisitor} from './utils/visitors.es';
 
 // Form/Data Engine Core
