+++ modules/apps/data-engine/data-engine-js-components-web/src/main/resources/META-INF/resources/js/utils/strings.js
